+++ app/Http/Controllers/Admin/DataSetController.php
@@ -320,8 +320,21 @@ class DataSetController extends AdminController
         }
 
         // prepare request for resources
+        $resPerPage = 10;
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+
         unset($params['dataset_uri']);
         $params['criteria']['dataset_uri'] = $uri;
+        $params['records_per_page'] = $resPerPage;
+        $params['page_number'] = $pageNumber;
+
+        if (isset($request->order)) {
+            $params['criteria']['order']['field'] = $request->order;
+        }
+
+        if (isset($request->order_type)) {
+            $params['criteria']['order']['type'] = $request->order_type;
+        }
 
         $resourcesReq = Request::create('/api/listResources', 'POST', $params);
         $apiResources = new ApiResource($resourcesReq);
@@ -364,12 +377,23 @@ class DataSetController extends AdminController
             return back();
         }
 
+        $paginationData = $this->getPaginationData(
+            $resources->resources,
+            $resources->total_records,
+            array_except(app('request')->input(), ['rpage']),
+            $resPerPage,
+            'rpage'
+        );
+
         return view('admin/datasetView', [
-            'class'     => 'user',
-            'dataset'   => $this->getModelUsernames($dataset),
-            'groups'    => $groups,
-            'setGroups' => $setGroups,
-            'resources' => $resources->resources,
+            'class'      => 'user',
+            'dataset'    => $this->getModelUsernames($dataset),
+            'groups'     => $groups,
+            'setGroups'  => $setGroups,
+            'resources'  => $paginationData['items'],
+            'pagination' => $paginationData['paginate'],
+            'uri'        => $uri,
+            'sorting'    => 'adminMyData'
         ]);
     }
 
@@ -561,11 +585,19 @@ class DataSetController extends AdminController
                     return redirect('/admin/resource/view/'. $response['uri']);
                 }
 
+                $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+                $resourcePaginationData = $this->getResourcePaginationData($response['data'], null, $pageNumber, true);
+
+                if (isset($response['data']['csvData'])) {
+                    $response['data']['csvData'] = $resourcePaginationData['data'];
+                }
+
                 return view('admin/resourceImport', array_merge([
                     'class'         => $class,
                     'types'         => $types,
                     'resourceUri'   => $response['uri'],
                     'action'        => 'create',
+                    'resPagination' => $resourcePaginationData['resPagination'],
                 ], $response['data']));
             } else {
                 // Delete resource record on fail
@@ -686,12 +718,16 @@ class DataSetController extends AdminController
             $versionsPerPage
         );
 
+        $pageNumber = !empty($request->rpage) ? $request->rpage : 1;
+        $resourcePaginationData = $this->getResourcePaginationData($data, $resource, $pageNumber);
+
         return view('admin/resourceView', [
             'class'         => 'user',
             'resource'      => $resource,
             'versions'      => $verData,
             'pagination'    => $paginationData['paginate'],
-            'data'          => $data,
+            'resPagination' => $resourcePaginationData['resPagination'],
+            'data'          => $resourcePaginationData['data'],
             'versionView'   => $version,
             'dataset'       => $datasetData,
             'supportName'   => !is_null($dataset) ? $dataset->support_name : null,