+++ resources/views/components/datasets/dataset_view.blade.php
@@ -167,6 +167,10 @@
             @endif
             <div class="col-xs-12 pull-left m-t-md p-l-none">
                 <div class="pull-left history">
+                    @if (!empty($resources) && count($resources) > 1)
+                        @include('partials.sorting.data-sets')
+                    @endif
+                    @include('partials.pagination')
                     @foreach ($resources as $resource)
                         @if (!empty($admin) || !empty($buttons[$resource->uri]['view']))
                             <div class="{{ $resource->reported ? 'signaled' : '' }}">
@@ -201,6 +205,7 @@
                             </div>
                         @endif
                     @endforeach
+                    @include('partials.pagination')
                 </div>
             </div>
             <div class="info-bar-sm col-sm-7 col-xs-12 p-l-none">