+++ resources/views/partials/resource-pagination.blade.php
@@ -1,0 +1,7 @@
+@if (isset($resPagination))
+    <div class="row">
+        <div class="col-xs-12 text-right">
+            {{ $resPagination->render() }}
+        </div>
+    </div>
+@endif